+++ apis/AccountApi.ts
@@ -62,7 +62,7 @@ export class AccountApi extends BaseApi {
         return AccountMapper.commercetoolsCustomerToAccount(response.body.customer);
       })
       .catch((error) => {
-        if (error.code === 400) {
+        if (error.statusCode === 400) {
           if (error instanceof ExternalError && error?.errorName === ExternalError.DUPLICATED_FIELD_ERROR_NAME) {
             throw new AccountEmailDuplicatedError({ message: `The account ${account.email} does already exist.` });
           }
@@ -75,7 +75,7 @@ export class AccountApi extends BaseApi {
           }
         }
 
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
 
     if (!account.confirmed) {
@@ -99,7 +99,7 @@ export class AccountApi extends BaseApi {
         return AccountMapper.commercetoolsCustomerToAccount(response.body);
       })
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 
@@ -132,7 +132,7 @@ export class AccountApi extends BaseApi {
         };
       })
       .catch((error) => {
-        if (error.code && error.code === 400) {
+        if (error.statusCode && error.statusCode === 400) {
           if (error.body && error.body?.errors?.[0]?.code === 'InvalidCredentials') {
             throw new AccountAuthenticationError({
               message: 'Failed to login account with the given credentials',
@@ -147,7 +147,7 @@ export class AccountApi extends BaseApi {
           }
         }
 
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
 
     if (!loggedInAccount.confirmed) {
@@ -176,7 +176,7 @@ export class AccountApi extends BaseApi {
         return AccountMapper.commercetoolsCustomerToAccount(response.body);
       })
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 
@@ -199,7 +199,7 @@ export class AccountApi extends BaseApi {
         };
       })
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 
@@ -218,7 +218,7 @@ export class AccountApi extends BaseApi {
         return AccountMapper.commercetoolsCustomerToAccount(response.body);
       })
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 
@@ -410,7 +410,7 @@ export class AccountApi extends BaseApi {
         return AccountMapper.commercetoolsCustomerToAccount(response.body);
       })
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 
@@ -433,7 +433,7 @@ export class AccountApi extends BaseApi {
         return AccountMapper.commercetoolsCustomerToAccount(response.body);
       })
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 
@@ -458,7 +458,7 @@ export class AccountApi extends BaseApi {
         return accountToken;
       })
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 
@@ -475,7 +475,7 @@ export class AccountApi extends BaseApi {
       .execute()
       .then((response) => response.body)
       .catch((error) => {
-        throw new ExternalError({ statusCode: error.code, message: error.message, body: error.body });
+        throw new ExternalError({ statusCode: error.statusCode, message: error.message, body: error.body });
       });
   }
 }