+++ interfaces/datasource.d.ts
@@ -1,0 +1,7 @@
+import '@frontastic/extension-types';
+
+declare module '@frontastic/extension-types' {
+  interface DataSourceConfiguration {
+    streamId?: string;
+  }
+}